hmmm... gcc 2.95.2 on AIX says sizeof is long...  fix up a printf to always have a match between the format string and argument
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@92339 13f79535-47bb-0310-9956-ffa450edef68

+++ server/main.c
@@ -89,7 +89,11 @@ static void show_compile_settings(void)
     printf("Server built:   %s\n", ap_get_server_built());
     printf("Server's Module Magic Number: %u:%u\n",
 	   MODULE_MAGIC_NUMBER_MAJOR, MODULE_MAGIC_NUMBER_MINOR);
-    printf("Architecture:   %d-bit\n", 8 * sizeof(void *));
+    /* sizeof(foo) is long on some platforms so we might as well
+     * make it long everywhere to keep the printf format
+     * consistent
+     */
+    printf("Architecture:   %ld-bit\n", 8 * (long)sizeof(void *));
     printf("Server compiled with....\n");
 #ifdef BIG_SECURITY_HOLE
     printf(" -D BIG_SECURITY_HOLE\n");
